Update React v1.1
